backend/requirements.txt
django-cors-headers
Django==3.2.19
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-tenant-schemas==1.11.0
django-mptt==0.14.0
psycopg2-binary==2.9.6
python-decouple==3.8
django-redis==5.2.0
pika==1.3.2
Pillow==9.5.0
gunicorn==21.2.0
redis==5.0.1